templates/fastapi-best-practice/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "FastAPI best-practice starter template (Mid — Postgres + Redis + JWT + Docker)"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    # FastAPI 코어
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic[email]>=2.9",
    "pydantic-settings>=2.6",
    # DB
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",        # alembic sync 드라이버
    # Cache / 외부 통신
    "redis>=5.0",
    "httpx>=0.27",
    # 로그
    "structlog>=24.4",
    # 인증
    "pyjwt>=2.10",
    "passlib[bcrypt]>=1.7",
    "bcrypt<4.1",                  # passlib 1.7.4 호환 (bcrypt 4.1+ 와 깨짐)
    "python-multipart>=0.0.18",    # OAuth2 form
    # 테스트 lifespan
    "asgi-lifespan>=2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
    "ruff>=0.7",
    "mypy>=1.13",
    "testcontainers[postgres,redis]>=4.8",
    "docker>=7.1",
    "types-passlib>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: testcontainers 가 필요한 통합 테스트 (도커 데몬 없으면 skip)",
]
